% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{conversion}
\alias{conversion}
\title{Conversion of units}
\usage{
conversion(source, target)
}
\arguments{
\item{source}{Source measurement unit. Currently available options are \code{"g/kg"}, \code{"\%"},
\code{"mg/kg"} and \code{"mg/dm^3"}.}

\item{target}{Target measurement unit. Currently available options are \code{"g/kg"} and \code{"mg/kg"}.}
}
\value{
A data.frame with source and target measurement units and their corresponding conversion factor.
}
\description{
Factors for converting measurement units of soil variables in the Brazilian Soil Iron Data Repository
(Fe-BR) -- \url{http://www.ufsm.br/febr}.
}
\details{
Target measurement units are defined according to the third edition of the Manual of Methods of Soil
Analysis published by Embrapa Soils in 2011.
}
\examples{
conversion(source = "\%", target = "g/kg")

}
\references{
Donagemma, G. K., Campos, D. V. B., Calderano, S. B., Teixeira, W. G. and Viana, J. H. M. (2011)
\emph{Manual de Métodos de Análise de Solo}. Rio de Janeiro: Embrapa Solos.
}
\seealso{
\url{http://www.ufsm.br/febr}
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
