% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset}
\alias{dataset}
\title{Get \emph{dataset} table}
\usage{
dataset(dataset, progress = TRUE, verbose = TRUE)
}
\arguments{
\item{dataset}{Character vector indicating one or more datasets. Identification codes should be as recorded
in \url{http://www.ufsm.br/febr/catalog/}. Use \code{dataset = "all"} to download all datasets.}

\item{progress}{(optional) Logical value indicating if a download progress bar should be displayed.}

\item{verbose}{(optional) Logical value indicating if informative messages should be displayed. Generally
useful to identify datasets with inconsistent data. Please report to \email{febr-forum@googlegroups.com}
if you find any issue.}
}
\value{
A list of data frames or a data frame with data of the chosen dataset(s).
}
\description{
Download data from the \emph{dataset} ("dataset") table of one or more datasets contained in the Free Brazilian
Repository for Open Soil Data -- \emph{\strong{febr}}, \url{http://www.ufsm.br/febr}. This includes dataset title and
description, author and institution identification, dataset license, and much more.
}
\examples{
\donttest{
res <- dataset(dataset = "ctb0003")
}
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
