% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasets}
\alias{datasets}
\title{Get soil datasets}
\usage{
datasets(progress = TRUE)
}
\arguments{
\item{progress}{Show progress bar?}
}
\value{
A list with all of the data of the soil datasets contained in Fe-BR.
}
\description{
Download dataset-specific data contained in the Brazilian Soil Iron Data Repository (Fe-BR).
}
\examples{
\dontrun{
res <- datasets()
str(res)
}
}
\seealso{
\url{http://www.ufsm.br/febr}
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
