% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{dictionary}
\alias{dictionary}
\alias{standard}
\title{Get \emph{standards} table}
\usage{
dictionary(table, variable, unit, precision, expr)

standard(table, variable, unit, precision, expr)
}
\arguments{
\item{table}{(optional) Character string indicating one or more table IDs that should be used for
filtering the dictionary. Accepted values: \code{"camada"}, \code{"observacao"}, \code{"metadado"}, and
\code{"versionamento"}.}

\item{variable}{(optional) Character vector indicating one or more variables. Accepts both specific
identification codes, e.g. \code{"ferro_oxalato_icpoes"} and \code{"carbono_cromo_30min150_mohr"}, as well as
general identification codes, e.g. \code{"ferro"} and \code{"carbono"}.}

\item{unit}{(optional) Character vector indicating one or more measurement units. For example, \code{"g/kg"},
\code{"g/cm^3"}, and \code{"cmolc/kg"}.}

\item{precision}{(optional) Integer vector indicating one or more number of decimal places.}

\item{expr}{(optional) Character string to be parsed and evaluated as a regular expression. For example,
\code{"campo_precisao > 0"}. Usage requires some knowledge of the structure of the \emph{standards} table.}
}
\value{
A data frame with standards for selected variable(s).
}
\description{
Download data from the \emph{standards} ("padroes") table of the Free Brazilian Repository for Open Soil Data
-- FEBR, \url{https://www.pedometria.org/febr/}. This includes codes and names of variable,
measurement units, number of decimal places, type of data, and description of analytical methods. This is
used to standardize the data contained in a dataset when downloading it via \code{\link[febr]{layer}} or
\code{\link[febr]{observation}}.
}
\examples{
#res <- standard(variable = "ferro")
#head(res)
}
\references{
Teixeira, P. C., Donagemma, G. K., Fontana, A., Teixeira, W. G. (2017)
\emph{Manual de Métodos de Análise de Solo}. Brasília: Embrapa.
}
\seealso{
The \emph{standards} table at
\url{https://docs.google.com/spreadsheets/d/1Dalqi5JbW4fg9oNkXw5TykZTA39pR5GezapVeV0lJZI}
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
