% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/febr-defunct.R
\name{dataset}
\alias{dataset}
\alias{febr}
\alias{febr2spdf}
\alias{febr2xlsx}
\alias{standard}
\alias{header}
\title{Defunct functions in the \strong{febr} package}
\usage{
dataset(...)
}
\arguments{
\item{...}{Not used.}
}
\description{
The functions listed here are no longer part of the \strong{febr} package as they are not needed (any
more).
}
