% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/febr2smartsolos.R
\name{febr2smartsolos}
\alias{febr2smartsolos}
\alias{smartsolos2febr}
\title{Conversion between FEBR and SMARTSolos soil profile data formats}
\usage{
febr2smartsolos(profiles, horizons, file, ...)

smartsolos2febr(file, ...)
}
\arguments{
\item{profiles}{Data frame with soil profile data, i.e. observation locations.}

\item{horizons}{Data frame with soil horizon data, i.e. sampling layers.}

\item{file}{Character string naming the JSON file to be read from or written to disk.}

\item{...}{(optional) Arguments passed to \code{\link[base]{writeLines}} and \code{\link[jsonlite]{fromJSON}}.}
}
\description{
Export FEBR soil profile data to the JSON file format required by the SMARTSolos API—and the other way
around.
}
\examples{
\donttest{
soil <- readFEBR("ctb0770", c("observacao", "camada"))[[1]]
febr2smartsolos(profiles = soil$observacao, horizons = soil$camada, file = "febr2smartsolos.json")
}
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
