% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{header}
\alias{header}
\title{Get table header}
\usage{
header(dataset, table, variable, stack = FALSE, progress = TRUE,
  verbose = TRUE)
}
\arguments{
\item{dataset}{Identification code of the dataset (or datasets) as recorded in 
\url{http://www.ufsm.br/febr/catalog/}. Use \code{dataset = "all"} to download all datasets.}

\item{table}{Table from which header data should be downloaded, with options \code{"camada"} 
(layer-specific header data) and \code{"observacao"} (observation-specific header data).}

\item{variable}{Name(s) of the variable(s). If missing, then a set of standard identification variables is
downloaded. Use \code{variable = "all"} to download all variables. See \sQuote{Details} for more information.}

\item{stack}{Should tables from different datasets be stacked on a single table for output? Defaults to
\code{stack = FALSE}, the output being a list of tables.}

\item{progress}{Show download progress bar?}

\item{verbose}{Show informative messages? Generally useful to identify datasets with inconsistent data. 
Please report to \email{febr-forum@googlegroups.com} if you find any issue.}
}
\value{
A list of data frames or a data frame with table header data (column names and measurement units) on
the chosen variable(s) of the chosen dataset(s).
}
\description{
Download header data (column names and measurement units) from table(s) containing layer or 
observation-specific data. This is useful to check what are the variables contained in a dataset before
downloading it via \code{\link[febr]{layer}} or \code{\link[febr]{observation}}.
}
\details{
\subsection{Standard columns}{
Standard columns and their content depend on the chosen \code{table}. See documentation of 
\code{\link[febr]{layer}} and \code{\link[febr]{observation}}.
}
}
\examples{
\dontrun{
res <- header(dataset = "all", table = "camada", variable = "ferro", stack = TRUE)
id <- grep("ferro_", colnames(res))
col <- colnames(res)[id]
col[order(col)]
}
}
\seealso{
\code{\link[febr]{layers}}, \code{\link[febr]{observations}}
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
