% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{identification}
\alias{identification}
\alias{dataset}
\title{Get 'identification' table}
\usage{
identification(data.set, progress = TRUE, verbose = TRUE, febr.repo = NULL)

dataset(data.set, progress = TRUE, verbose = TRUE, febr.repo = NULL)
}
\arguments{
\item{data.set}{Character vector indicating the identification code of one or more data sets.
Use \code{data.set = "all"} to download all data sets.}

\item{progress}{(optional) Logical value indicating if a download progress bar should be displayed.}

\item{verbose}{(optional) Logical value indicating if informative messages should be displayed. Generally
useful to identify datasets with inconsistent data. Please report to \email{febr-forum@googlegroups.com}
if you find any issue.}

\item{febr.repo}{(optional) Defaults to the remote file directory of the Federal University of
Technology - Paraná at \url{https://cloud.utfpr.edu.br/index.php/s/Df6dhfzYJ1DDeso}.
Alternatively, a local directory path can be informed if the user has a local copy of the data
repository.}

\item{dataset}{Character vector indicating the identification code of one or more data sets.
Use \code{dataset = "all"} to download all data sets.}
}
\value{
A list of data frames or a data frame with data of the chosen dataset(s).
}
\description{
Download data from the 'identification' ("identificacao") table of one or more soil datasets
published in the Free Brazilian Repository for Open Soil Data (FEBR),
\url{https://www.pedometria.org/febr/}. This table includes data such as dataset title and
description, author and institution, data license, and much more.
}
\note{
Check the new core data download function \code{readFEBR()}.
}
\examples{
\donttest{
res <- identification(data.set = "ctb0003")
}
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
