% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{layer}
\alias{layer}
\title{Get layer data}
\usage{
layer(dataset, variable, stack = FALSE, missing = list(depth = "keep", data
  = "keep"), standardization = list(plus.sign = "keep", plus.depth = 2.5,
  transition = "keep", smoothing.fun = "mean", units = FALSE, round = FALSE),
  harmonization = list(harmonize = FALSE, level = 2), progress = TRUE,
  verbose = TRUE)
}
\arguments{
\item{dataset}{Identification code of the dataset (or datasets) as recorded in 
\url{http://www.ufsm.br/febr/catalog/}. Use \code{dataset = "all"} to download all datasets.}

\item{variable}{Name(s) of the variable(s). If missing, then a set of standard identification variables is
downloaded. Use \code{variable = "all"} to download all variables. See \sQuote{Details} for more information.}

\item{stack}{Should tables from different datasets be stacked on a single table for output? Defaults to
\code{stack = FALSE}, the output being a list of tables.}

\item{missing}{List with named sub-arguments specifying what should be done with a layer missing data on
sampling depth, \code{depth}, or data on variable(s), \code{data}? Options are \code{"keep"} (default) and \code{"drop"}.}

\item{standardization}{List named sub-arguments specifying how to perform data standardization.
\itemize{
\item \code{plus.sign} What should be done with the plus sign (\code{+}) commonly used along with the inferior
limit of the bottom layer of an observation? Options are \code{"keep"} (default), \code{"add"}, and \code{"remove"}.
\item \code{plus.depth} Depth increment (in centimetres) when processing the plus sign (\code{+}) with
\code{plus.sign = "add"}. Defaults to \code{plus.depth = 2.5}.
\item \code{transition} What should be done about wavy, irregular, and broken transitions between layers in an
observation? Options are \code{"keep"} (default) and \code{"smooth"}.
\item \code{smoothing.fun} Function that should be used to smooth wavy and irregular transitions between
layers in an observation when \code{transition = "smooth"}. Options are \code{"mean"} (default), \code{"min"},
\code{"max"}, and \code{"median"}.
\item \code{units} Should the values of the real and integer variable(s) be converted to the standard
measurement unit(s)? Defaults to \code{units = FALSE}, i.e. no conversion is performed.
\item \code{round} Should the values of the real and integer variable(s) be rounded to the standard number
of decimal places? Effective only when \code{units = TRUE}. Defaults to \code{round = FALSE}, i.e.
no rounding is performed.
}}

\item{harmonization}{List with named sub-arguments specifying if and how to perform data harmonization.
\itemize{
\item \code{harmonize} Should data be harmonized? Defaults to \code{harmonize = FALSE}, i.e. no
harmonization is performed.
\item \code{level} Number of levels of the identification code of the variables that should be considered
for harmonization. Defaults to \code{level = 2}. See \sQuote{Details} for more information.
}}

\item{progress}{Show download progress bar?}

\item{verbose}{Show informative messages? Generally useful to identify datasets with inconsistent data. 
Please report to \email{febr-forum@googlegroups.com} if you find any issue.}
}
\value{
A list of data frames or a data frame with layer-specific data on the chosen variable(s) of
the chosen dataset(s).
}
\description{
Download layer-specific data (sampling depth, layer designation, among others) contained in the
Free Brazilian Repository for Open Soil Data -- \emph{\strong{febr}}, \url{http://www.ufsm.br/febr}. In \emph{\strong{febr}},
layer-specific data are stored using a table named \code{"camada"}. Use \code{\link[febr]{header}} if you
want to check what are the layer-specific variables contained in a dataset before downloading it.
}
\details{
\subsection{Standard columns}{
Standard columns and their content (in Portuguese) are as follows:
\itemize{
\item \code{dataset_id}. Identification code of the dataset in \emph{\strong{febr}} to which an observation belongs.
\item \code{observacao_id}. Identification code of an observation in \emph{\strong{febr}}.
\item \code{camada_numero}. Sequential layer number, from top to bottom.
\item \code{camada_nome}. Layer designation according to some standard description guide.
\item \code{amostra_codigo}. Laboratory number of a sample.
\item \code{profund_sup}. Upper boundary of a layer (cm).
\item \code{profund_inf}. Lower boundary of a layer (cm).
}
Further details about the content of the standard columns can be found in \url{http://www.ufsm.br/febr/book/}
(in Portuguese).
}

\subsection{Harmonization}{
Data harmonization consists of converting the values of a variable determined using some method \emph{B} so
that they are (approximately) equivalent to the values that would have been obtained if the standard method
\emph{A} had been used instead. For example, converting carbon content values obtained using a wet digestion
method to the standard dry combustion method is data harmonization.

A heuristic data harmonization procedure is implemented \strong{febr}. It consists of grouping variables
based on a chosen number of levels of their identification code. For example, consider a variable with an
identification code composed of four levels, \code{aaa_bbb_ccc_ddd}, where \code{aaa} is the first level and
\code{ddd} is the fouth level. Now consider a related variable, \code{aaa_bbb_eee_fff}. If the harmonization
is to consider all four coding levels (\code{level = 4}), then these two variables will remain coded as
separate variables. But if \code{level = 2}, then both variables will be recoded to \code{aaa_bbb}, thus becoming the
same variable.
}
}
\examples{
\dontrun{
res <- layer(dataset = paste("ctb000", 4:9, sep = ""))
str(res)
}
}
\seealso{
\code{\link[febr]{observations}}, \code{\link[febr]{standards}}, \code{\link[febr]{conversion}}
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
