% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layers}
\alias{layers}
\title{Get soil layers}
\usage{
layers(which.cols = "standard", progress = TRUE)
}
\arguments{
\item{which.cols}{Which columns should be returned? Options are \code{"standard"} (default) and
\code{"all"}.}

\item{progress}{Show progress bar?}
}
\value{
A list with some or all of the data of the soil layers contained in Fe-BR.
}
\description{
Download layer-specific data contained in the Brazilian Soil Iron Data Repository (Fe-BR).
}
\details{
Standard columns are as follows:
\itemize{
\item \code{dataset_id}. Identification code of the respective dataset.
\item \code{observation_id}. Identification code of the respective observation.
\item \code{layer_number}. Sequential layer number, from top to bottom.
\item \code{layer_name}. Layer designation according to some standard soil description guidelines.
\item \code{sample_code}. Laboratory number of the soil sample.
\item \code{upper_depth}. Limite superior da camada.
\item \code{lower_depth}. Limite inferior da camada.
\item \code{fe_xxx_yyy}. Soil iron content data, with \code{xxx} being a given extractant and \code{yyy}
being a given determination technique/equipment.
}
}
\examples{
\dontrun{
res <- layers()
str(res)
}
}
\seealso{
\url{http://www.ufsm.br/febr}
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
