% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\title{Get soil metadata}
\usage{
metadata(progress = TRUE)
}
\arguments{
\item{progress}{Show progress bar?}
}
\value{
A list with metadata-specific data.
}
\description{
Download metadata-specific data contained in the Brazilian Soil Iron Data Repository  --
\url{http://www.ufsm.br/febr}.
}
\examples{
\dontrun{
res <- metadata()
str(res)
}
}
\seealso{
\url{http://www.ufsm.br/febr}
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
