% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{observations}
\alias{observations}
\title{Get soil observations}
\usage{
observations(which.cols = "standard", stack.obs = TRUE,
  missing.coords = "drop", progress = TRUE)
}
\arguments{
\item{which.cols}{Which columns should be returned? Options are \code{"standard"} (default) and
\code{"all"}.}

\item{stack.obs}{Should observations from different datasets be stacked on a single data frame for output?
Used only with \code{which.cols = "standard"}. Defaults to \code{stack.obs = TRUE}.}

\item{missing.coords}{What should be done with observations missin spatial coordinates? Options are
\code{"drop"} (default) and \code{"keep"}.}

\item{progress}{Show progress bar?}
}
\value{
A list or data.frame with some or all of the data of the soil observations contained in Fe-BR.
}
\description{
Download observation-specific data contained in the Brazilian Soil Iron Data Repository (Fe-BR).
}
\examples{
\dontrun{
res <- observations(which.cols = "standard", stack.obs = TRUE)
str(res)
}
}
\seealso{
www.ufsm.br/febr
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
