% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readIndex}
\alias{readIndex}
\title{Utilities}
\usage{
readIndex(...)
}
\arguments{
\item{...}{(optional) Arguments passed to \code{\link[data.table]{fread}}.}
}
\description{
Utilities
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
