% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization.R
\name{depth}
\alias{depth}
\title{Data standardization}
\usage{
depth(layers, plus.sign = c("remove", "add"), plus.depth)
}
\arguments{
\item{layers}{Data frame with soil horizon data, i.e. sampling layers.}

\item{plus.sign}{plus.sign}

\item{plus.depth}{plus.depth}
}
\description{
Data standardization
}
