% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standards.R
\name{standards}
\alias{standards}
\title{Standards for soil variables}
\usage{
standards(soil.var = "fe", extraction.method)
}
\arguments{
\item{soil.var}{Identification code of the soil variable. The only currently available option is 
\code{soil.var = "fe"}.}

\item{extraction.method}{Identification code of the extraction method. See \sQuote{Details}.}
}
\value{
A data.frame.
}
\description{
Definitions of standards for soil variables in the Free Brazilian Repository for Open Soil Data -- 
\url{http://www.ufsm.br/febr}.
}
\details{
Standards for soil variables are defined according to the third edition of the Manual of Methods of Soil
Analysis published by Embrapa Soils in 2011. They include the measurement units and number of digital 
places that should be used with a soil variable as defined by the extraction method.
}
\examples{
# All standards
standards()

# Specific stantdard
standards(extraction.method = "sulfurico")

}
\references{
Donagemma, G. K., Campos, D. V. B., Calderano, S. B., Teixeira, W. G. and Viana, J. H. M. (2011)
\emph{Manual de Métodos de Análise de Solo}. Rio de Janeiro: Embrapa Solos.
}
\seealso{
\url{http://www.ufsm.br/febr}
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
