% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasets}
\alias{datasets}
\title{Get soil datasets}
\usage{
datasets(dataset, progress = TRUE, verbose = TRUE)
}
\arguments{
\item{dataset}{Identification code of the dataset (or datasets). Use \code{dataset = "all"} to download 
all datasets.}

\item{progress}{Show download progress bar?}

\item{verbose}{Show informative messages? Generally useful to identify datasets with inconsistent data. 
Please report to \email{febr-forum@googlegroups.com} if you find any issue.}
}
\value{
A list with dataset-specific data.
}
\description{
Download soil dataset-specific data contained in the Free Brazilian Repository for Open Soil Data --
\url{http://www.ufsm.br/febr}.
}
\examples{
\dontrun{
res <- datasets()
str(res)
}
}
\seealso{
\url{http://www.ufsm.br/febr}
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
