% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/febr2sse.R
\name{febr2sse}
\alias{febr2sse}
\title{Conversion between FEBR and SmartSolos Expert (SSE) soil profile data formats}
\usage{
febr2sse(profiles, horizons, file, ...)
}
\arguments{
\item{profiles}{Data frame with soil profile data, i.e. observation locations.}

\item{horizons}{Data frame with soil horizon data, i.e. sampling layers.}

\item{file}{(optional) Character string naming the JSON file to be read from or written to disk.}

\item{...}{(optional) Arguments passed to \code{\link[base]{writeLines}} and
\code{\link[jsonlite]{fromJSON}}.}
}
\description{
Export FEBR soil profile data to the JSON file format required by the SmartSolos Expert API.
}
\examples{
\donttest{
profiles <- observation(
  dataset = "ctb0025", variable = c("taxon_sibcs", "relevo_drenagem"),
  standardization = list(units = TRUE, round = TRUE))
idx <- profiles$observacao_id[1]
profiles <- profiles[profiles$observacao_id \%in\% idx, ]
horizons <- layer(
  dataset = "ctb0025", variable = "all",
  standardization =
    list(plus.sign = "remove", lessthan.sign = "remove",
         transition = "smooth", units = TRUE, round = TRUE))
horizons <- horizons[horizons$observacao_id \%in\% idx, ]
horizons[, 7:46] <- lapply(horizons[, 7:46], as.numeric)
horizons <- cbind(
  horizons,
  morphology(x = horizons$morfologia_descricao, variable = "color"),
  morphology(x = horizons$morfologia_descricao, variable = "structure"),
  morphology(x = horizons$morfologia_descricao, variable = "consistence"),
  stringsAsFactors = FALSE)
file <- ifelse(
  dir.exists("tmp"),
  paste0("tmp/febr2smartsolos-", idx, ".json"),
  paste0("febr2smartsolos-", idx, ".json"))
febr2sse(profiles, horizons, file)
}
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
