% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goto.R
\name{goto}
\alias{goto}
\title{Go to FEBR}
\usage{
goto(dataset, page)
}
\arguments{
\item{dataset}{(optional) Character vector indicating one dataset.}

\item{page}{(optional) Character string indicating a web page of the FEBR. Options are: \code{"febr"},
\code{"search"}, \code{"package"}, \code{"github"}, \code{"forum"}, \code{"units"}, \code{"standards"}, \code{"index"}, and \code{"template"}.}
}
\description{
Go to one of the web pages of the Free Brazilian Repository for Open Soil Data -- FEBR --,
including project and dataset web pages.
}
\examples{
\donttest{
# Go to the main project page
goto(page = "febr")
}
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
