% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{layer}
\alias{layer}
\title{Get layer data}
\usage{
layer(dataset, variable, stack = FALSE, missing = list(depth = "keep", data
  = "keep"), standardization = list(plus.sign = "keep", plus.depth = 0,
  transition = "keep", smoothing.fun = "mean"), harmonization = list(level =
  5), progress = TRUE, verbose = TRUE)
}
\arguments{
\item{dataset}{Identification code of the dataset (or datasets) as recorded in 
\url{http://www.ufsm.br/febr/catalog/}. Use \code{dataset = "all"} to download all datasets.}

\item{variable}{Name(s) of the variable(s). If missing, then a set of standard identification variables is
downloaded. Use \code{variable = "all"} to download all variables. See \sQuote{Details} for more information.}

\item{stack}{Should tables from different datasets be stacked on a single table for output? Defaults to
\code{stack = FALSE}, the output being a list of tables.}

\item{missing}{List with named sub-arguments specifying what should be done with layers missing depth, 
\code{depth}, or data on variables, \code{data}? Options are \code{"keep"} (default) and \code{"drop"}.}

\item{standardization}{List with definitions on how to \emph{standardize} layer-specific data.
\itemize{
\item \code{plus.sign} What should be done with the plus sign ('+') commonly used along with the inferior 
      limit of the bottom layer of an observation? Options are \code{"keep"} (default), \code{"add"},
      and \code{"remove"}.
\item \code{plus.depth} Depth increment (in centimetres) when processing the plus sign ('+') with 
      \code{plus.sign = "add"}. Defaults to \code{plus.depth = 0}.
\item \code{transition} What should be done about wavy, irregular, and broken transitions between layers in
      an observation? Options are \code{"keep"} (default) and \code{"smooth"}.
\item \code{smoothing.fun} Function that should be used to smooth wavy and irregular transitions between 
      layers in an observation when \code{transition = "smooth"}. Options are \code{"mean"} (default),
      \code{"min"}, \code{"max"}, and \code{"median"}. 
}}

\item{harmonization}{List with definitions on how to \emph{harmonize} layer-specific data.
\itemize{
\item \code{level} Level of harmonization. Defaults to \code{level = 5}. See \code{\link[febr]{standards}}.
}}

\item{progress}{Show download progress bar?}

\item{verbose}{Show informative messages? Generally useful to identify datasets with inconsistent data. 
Please report to \email{febr-forum@googlegroups.com} if you find any issue.}
}
\value{
A list of data frames or a data frame with layer-specific data on the chosen variable(s) of 
the chosen dataset(s).
}
\description{
Download layer-specific data (sampling depth, layer designation, among others) contained in the
Free Brazilian Repository for Open Soil Data -- \pkg{febr}, \url{http://www.ufsm.br/febr}. In \pkg{febr},
layer-specific data are stored using a table named \code{"camada"}. Use \code{\link[febr]{header}} if you
want to check what are the variables contained in a dataset before downloading it.
}
\details{
\subsection{Standard columns}{
Standard columns and their content are as follows:
\itemize{
\item \code{dataset_id}. Identification code of the dataset in \pkg{febr} to which an observation belongs.
\item \code{observacao_id}. Identification code of an observation in \pkg{febr}.
\item \code{camada_numero}. Sequential layer number, from top to bottom.
\item \code{camada_nome}. Layer designation according to some standard description guide.
\item \code{amostra_codigo}. Laboratory number of a sample.
\item \code{profund_sup}. Upper boundary of a layer (cm).
\item \code{profund_inf}. Lower boundary of a layer (cm).
}
Further details about the content of the standard columns can be found in \url{http://www.ufsm.br/febr/book/}
(in Portuguese).
}
}
\examples{
\dontrun{
res <- layer(dataset = paste("ctb000", 4:9, sep = ""))
str(res)
}
}
\seealso{
\code{\link[febr]{observations}}, \code{\link[febr]{standards}}, \code{\link[febr]{conversion}}
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
