% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layers}
\alias{layers}
\title{Get soil layers}
\usage{
layers(dataset = "all", which.cols = "standard", soil.vars = "fe",
  stack.datasets = TRUE, missing.data = "drop",
  standardization = list(plus.sign = "add", plus.depth = 2.5, transition =
  "smooth", smoothing.fun = "mean"), harmonization = list(level = 1),
  progress = TRUE, verbose = TRUE)
}
\arguments{
\item{dataset}{Identification code of the dataset or datasets for which soil layer-specific data should be
downloaded -- see \url{http://www.ufsm.br/febr/data}. Defaults to \code{dataset = "all"}, that is, download
data from all existing datasets.}

\item{which.cols}{Which columns should be returned? Options are \code{"standard"} (default) and
\code{"all"}. See \sQuote{Details} for a description of the standard columns.}

\item{soil.vars}{Identification code of the soil variables for which soil layer-specific data should be
downloaded. The only currently available option is \code{soil.vars = "fe"}, that is, various forms of soil 
iron content.}

\item{stack.datasets}{Should soil layers from different datasets be stacked on a single data frame for
output? Used only with \code{which.cols = "standard"}. Defaults to \code{stack.datasets = TRUE}.}

\item{missing.data}{What should be done with soil layers missing any iron data? Options are \code{"drop"}
(default) and \code{"keep"}.}

\item{standardization}{List with definitions on how to \emph{standardize} soil layer-specific data.
\itemize{
\item \code{plus.sign} What should be done with the plus sign ('+') commonly used along with the inferior 
      limit of the bottom layer of soil observations? Options are \code{"add"} (default), \code{"remove"},
      and \code{"keep"}.
\item \code{plus.depth} Depth increment (in centimetres) when processing the plus sign ('+') with 
      \code{plus.sign = "add"}. Defaults to \code{plus.depth = 2.5}.
\item \code{transition} What should be done about wavy, irregular, and broken transitions between layers in
      a soil observation? Options are \code{"smooth"} (default) and \code{"keep"}.
\item \code{smoothing.fun} Function that should be used to smooth wavy and irregular transitions between 
      layers in a soil observation when \code{transition = "smooth"}. Options are \code{"mean"} (default),
      \code{"min"}, \code{"max"}, and \code{"median"}. 
}}

\item{harmonization}{List with definitions on how to \emph{harmonize} soil layer-specific data.
\itemize{
\item \code{level} Should data on soil variables be harmonized based only on the extraction method, 
      \code{level = 1} (default), or on both extraction and measurement methods, \code{level = 2}? See 
      \code{\link[febr]{standards}}.
}}

\item{progress}{Show download progress bar?}

\item{verbose}{Show informative messages? Generally useful identify datasets with any inconsistent data. 
Please report to \email{fe-br@googlegroups.com} if you find any issue.}
}
\value{
A list or data.frame with soil layer-specific data.
}
\description{
Download soil layer-specific data contained in the Free Brazilian Repository for Open Soil Data --
\url{http://www.ufsm.br/febr}.
}
\details{
\subsection{Standard columns}{
Standard columns and their content are as follows:
\itemize{
\item \code{dataset_id}. Identification code of the datasets in febr to which soil observations belong.
\item \code{observacao_id}. Identification code of soil observations in febr.
\item \code{camada_numero}. Sequential layer number, from top to bottom.
\item \code{camada_nome}. Layer designation according to some standard soil description guidelines.
\item \code{amostra_codigo}. Laboratory number of the soil samples.
\item \code{profund_sup}. Upper boundary of soil layers (cm).
\item \code{profund_inf}. Lower boundary of soil layers (cm).
\item \code{fe_xxx_yyy}. Soil iron content data, with \code{xxx} being a given extraction method and
\code{yyy} being a given measurement method.
}
}
}
\examples{
\dontrun{
res <- layers(dataset = paste("ctb000", 4:9, sep = ""))
str(res)
}
}
\seealso{
\code{\link[febr]{observations}}, \code{\link[febr]{standards}}, \code{\link[febr]{conversion}}
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
