% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\title{Get soil metadata}
\usage{
metadata(dataset, progress = TRUE)
}
\arguments{
\item{dataset}{Identification code of the dataset (or datasets) for which soil metadata-specific data should
be downloaded -- see \url{http://www.ufsm.br/febr/book}. Use \code{dataset = "all"} to download data from 
all existing datasets.}

\item{progress}{Show progress bar?}
}
\value{
A list with metadata-specific data.
}
\description{
Download metadata-specific data contained in the Free Brazilian Repository for Open Soil Data --
\url{http://www.ufsm.br/febr}.
}
\examples{
\dontrun{
res <- metadata()
str(res)
}
}
\seealso{
\url{http://www.ufsm.br/febr}
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
