% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation.R
\name{observation}
\alias{observation}
\title{Get observation data}
\usage{
observation(dataset, variable, stack = FALSE, missing = list(coord = "keep",
  data = "keep"), standardization = list(crs = NULL, units = FALSE, round =
  FALSE), harmonization = list(harmonize = FALSE, level = 2),
  progress = TRUE, verbose = TRUE)
}
\arguments{
\item{dataset}{Identification code of the dataset (or datasets) as recorded in 
\url{http://www.ufsm.br/febr/catalog/}. Use \code{dataset = "all"} to download all datasets.}

\item{variable}{Name(s) of the variable(s). If missing, then a set of standard identification variables is
downloaded. Use \code{variable = "all"} to download all variables. See \sQuote{Details} for more information.}

\item{stack}{Should tables from different datasets be stacked on a single table for output? Defaults to
\code{stack = FALSE}, the output being a list of tables.}

\item{missing}{List with named sub-arguments specifying what should be done with an observation missing
spatial coordinates, \code{coord}, or data on variables, \code{data}? Options are \code{"keep"} (default)
and \code{"drop"}.}

\item{standardization}{List named sub-arguments specifying how to perform data standardization.
\itemize{
\item \code{crs} EPSG code defining the target coordinate reference system (CRS) to which spatial
coordinates should be transformed. For example, \code{crs = "EPSG:4674"}, i.e. SIRGAS 2000, the
standard CRS for Brazil -- see more at \url{http://spatialreference.org/ref/epsg/}. Defaults to
\code{crs = NULL}, i.e. no transformation is performed.
\item \code{units} Should the values of the real and integer variable(s) be converted to the standard
measurement unit(s)? Defaults to \code{units = FALSE}, i.e. no conversion is performed.
\item \code{round} Should the values of the real and integer variable(s) be rounded to the standard number
of decimal places? Effective only when \code{units = TRUE}. Defaults to \code{round = FALSE}, i.e.
no rounding is performed.
}}

\item{harmonization}{List with named sub-arguments specifying if and how to perform data harmonization.
\itemize{
\item \code{harmonize} Should data be harmonized? Defaults to \code{harmonize = FALSE}, i.e. no
harmonization is performed.
\item \code{level} Number of levels of the identification code of the variables that should be considered
for harmonization. Defaults to \code{level = 2}. See \sQuote{Details} for more information.
}}

\item{progress}{Show download progress bar?}

\item{verbose}{Show informative messages? Generally useful to identify datasets with inconsistent data. 
Please report to \email{febr-forum@googlegroups.com} if you find any issue.}
}
\value{
A list of data frames or a data frame with observation-specific data on the chosen variable(s) of
the chosen dataset(s).
}
\description{
Download observation-specific data (spatial coordinates, observation date, among others) contained in the
Free Brazilian Repository for Open Soil Data -- \pkg{febr}, \url{http://www.ufsm.br/febr}. In \pkg{febr},
observation-specific data are stored using a table named \code{"observacao"}. Use \code{\link[febr]{header}}
if you want to check what are the variables contained in a dataset before downloading it.
}
\details{
\subsection{Standard columns}{
Standard columns and their content (in Portuguese) are as follows:
\itemize{
\item \code{dataset_id}. Identification code of the dataset in \pkg{febr} to which an observation belongs.
\item \code{observacao_id}. Identification code of an observation in \pkg{febr}.
\item \code{sisb_id}. Identification code of an observation in the Brazilian Soil Information System
maintained by the Brazilian Agricultural Research Corporation (EMBRAPA) at
\url{https://www.bdsolos.cnptia.embrapa.br/consulta_publica.html}.
\item \code{ibge_id}. Identification code of an observation in the database of the Brazilian Institute
of Geography and Statistics (IBGE) at \url{http://www.downloads.ibge.gov.br/downloads_geociencias.htm#}.
\item \code{observacao_data}. Date (dd-mm-yyyy) in which an observation was made.
\item \code{coord_sistema}. EPSG code of the coordinate reference system.
\item \code{coord_x}. Longitude (°) or Easting (m).
\item \code{coord_y}. Latitude (°) or Northing (m).
\item \code{coord_precisao}. Precision with which x- and y-coordinates were determined (m).
\item \code{coord_fonte}. Source of the x- and y-coordinates.
\item \code{pais_id}. Country code (ISO 3166-1 alpha-2).
\item \code{estado_id}. Code of the Brazilian federative unit where an observation was made.
\item \code{municipio_id}. Name of the Brazilian county where as observation was made.
\item \code{amostra_tipo}. Type of sample taken.
\item \code{amostra_quanti}. Number of samples taken.
\item \code{amostra_area}. Sampling area.
}
Further details about the content of the standard columns can be found in \url{http://www.ufsm.br/febr/book/}
(in Portuguese).
}

\subsection{Harmonization}{
Data harmonization consists of converting the values of a variable determined using some method \emph{B} so
that they are (approximately) equivalent to the values that would have been obtained if the standard method
\emph{A} had been used instead. For example, converting carbon content values obtained using a wet digestion
method to the standard dry combustion method is data harmonization.

A heuristic data harmonization procedure is implemented \strong{febr}. It consists of grouping variables
based on a chosen number of levels of their identification code. For example, consider a variable with an
identification code composed of four levels, \code{aaa_bbb_ccc_ddd}, where \code{aaa} is the first level and
\code{ddd} is the fouth level. Now consider a related variable, \code{aaa_bbb_eee_fff}. If the harmonization
is to consider all four coding levels (\code{level = 4}), then these two variables will remain coded as
separate variables. But if \code{level = 2}, then both variables will be recoded to \code{aaa_bbb}, thus becoming the
same variable.
}
}
\examples{
\dontrun{
res <- observation(dataset = paste("ctb000", 4:9, sep = ""), variable = "taxon")
str(res)
}
}
\seealso{
\code{\link[febr]{layers}}
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
