% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{observations}
\alias{observations}
\title{Get soil observations}
\usage{
observations(dataset = "all", which.cols = "standard", stack.obs = TRUE,
  missing.coords = "drop", target.crs = "EPSG:4674", progress = TRUE)
}
\arguments{
\item{dataset}{Character vector with the identification code of the dataset -- or datasets -- for which 
soil observation-specific data should be downloaded. Defaults to \code{dataset = "all"}, that is, download
data from all existing datasets.}

\item{which.cols}{Which columns should be returned? Options are \code{"standard"} (default) and
\code{"all"}.}

\item{stack.obs}{Should soil observations from different datasets be stacked on a single data frame for
output? Used only with \code{which.cols = "standard"}. Defaults to \code{stack.obs = TRUE}.}

\item{missing.coords}{What should be done with soil observations missing spatial coordinates? Options are
\code{"drop"} (default) and \code{"keep"}.}

\item{target.crs}{EPSG code defining the target coordinate reference system to which spatial coordinates
should be transformed. Used only with \code{stack.obs = TRUE}. Defaults to \code{target.crs = "EPSG:4674"},
i.e. SIRGAS 2000, the standard CRS for Brazil -- see more at \url{http://spatialreference.org/ref/epsg/}.
If set to \code{target.crs = NULL} then the native spatial coordinates are returned.}

\item{progress}{Show download progress bar?}
}
\value{
A list or data.frame with soil observation-specific data.
}
\description{
Download soil observation-specific data contained in the Brazilian Soil Iron Data Repository (Fe-BR) --
\url{http://www.ufsm.br/febr}.
}
\details{
Standard columns and their content are as follows:
\itemize{
\item \code{dataset_id}. Identification code of the datasets in Fe-BR to which soil observations belong.
\item \code{observacao_id}. Identification code of soil observations in Fe-BR.
\item \code{sisb_id}. Identification code of soil observations in the Brazilian Soil Information System
maintained by the Brazilian Agricultural Research Corporation (EMBRAPA) at
\url{https://www.bdsolos.cnptia.embrapa.br/consulta_publica.html}.
\item \code{ibge_id}. Identification code of soil observations in the database of the Brazilian Institute
of Geography and Statistics (IBGE) at \url{http://www.downloads.ibge.gov.br/downloads_geociencias.htm#}.
\item \code{observacao_data}. Date (dd-mm-yyyy) in which soil observations were made.
\item \code{coord_sistema}. EPSG code of the coordinate reference system -- see more at
\url{http://spatialreference.org/ref/epsg/}.
\item \code{coord_x}. Longitude (°) or Easting (m).
\item \code{coord_y}. Latitude (°) or Northing (m).
\item \code{coord_precisao}. Precision with which x- and y-coordinates were determined (m).
\item \code{coord_fonte}. Source of the x- and y-coordinates.
\item \code{pais_id}. Country code (ISO 3166-1 alpha-2), i.e. \code{"BR"}.
\item \code{estado_id}. Code of the Brazilian federative units.
\item \code{municipio_id}. Name of the Brazilian county where soil observations were made.
\item \code{amostra_tipo}. Type of soil sample taken, i.e. simple or composed.
\item \code{amostra_quanti}. Number of soil samples taken.
\item \code{amostra_area}. Sampling area.
}
}
\examples{
\dontrun{
res <- observations(dataset = "ctb0016")
str(res)
}
}
\seealso{
\url{http://www.ufsm.br/febr}
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
