% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{observations}
\alias{observations}
\title{Get soil observations}
\usage{
observations(dataset, which.cols = "standard", stack.obs = TRUE,
  missing.coords = "drop", target.crs = "EPSG:4674", progress = TRUE,
  verbose = TRUE)
}
\arguments{
\item{dataset}{Identification code of the dataset (or datasets) for which soil layer-specific data should be
downloaded -- see \url{http://www.ufsm.br/febr/book}. Use \code{dataset = "all"} to download data from all
existing datasets.}

\item{which.cols}{Which columns should be returned? Options are \code{"standard"} (default) and
\code{"all"}. See \sQuote{Details} for a description of the standard columns.}

\item{stack.obs}{Should soil observations from different datasets be stacked on a single data frame for
output? Used only with \code{which.cols = "standard"}. Defaults to \code{stack.obs = TRUE}.}

\item{missing.coords}{What should be done with soil observations missing spatial coordinates? Options are
\code{"drop"} (default) and \code{"keep"}.}

\item{target.crs}{EPSG code defining the target coordinate reference system to which spatial coordinates
should be transformed. Defaults to \code{target.crs = "EPSG:4674"}, i.e. SIRGAS 2000, the standard CRS for
Brazil -- see more at \url{http://spatialreference.org/ref/epsg/}. If set to \code{target.crs = NULL} then
the native spatial coordinates are returned.}

\item{progress}{Show download progress bar?}

\item{verbose}{Show informative messages? Generally useful to identify datasets with any inconsistent data. 
Please report to \email{febr-forum@googlegroups.com} if you find any issue.}
}
\value{
A list or data.frame with soil observation-specific data.
}
\description{
Download soil observation-specific data contained in the Free Brazilian Repository for Open Soil Data --
\url{http://www.ufsm.br/febr}.
}
\details{
\subsection{Standard columns}{
Standard columns and their content are as follows:
\itemize{
\item \code{dataset_id}. Identification code of the datasets in febr to which soil observations belong.
\item \code{observacao_id}. Identification code of soil observations in febr.
\item \code{sisb_id}. Identification code of soil observations in the Brazilian Soil Information System
maintained by the Brazilian Agricultural Research Corporation (EMBRAPA) at
\url{https://www.bdsolos.cnptia.embrapa.br/consulta_publica.html}.
\item \code{ibge_id}. Identification code of soil observations in the database of the Brazilian Institute
of Geography and Statistics (IBGE) at \url{http://www.downloads.ibge.gov.br/downloads_geociencias.htm#}.
\item \code{observacao_data}. Date (dd-mm-yyyy) in which soil observations were made.
\item \code{coord_sistema}. EPSG code of the coordinate reference system -- see more at
\url{http://spatialreference.org/ref/epsg/}.
\item \code{coord_x}. Longitude (°) or Easting (m).
\item \code{coord_y}. Latitude (°) or Northing (m).
\item \code{coord_precisao}. Precision with which x- and y-coordinates were determined (m).
\item \code{coord_fonte}. Source of the x- and y-coordinates.
\item \code{pais_id}. Country code (ISO 3166-1 alpha-2), i.e. \code{"BR"}.
\item \code{estado_id}. Code of the Brazilian federative unit where soil observations were made.
\item \code{municipio_id}. Name of the Brazilian county where soil observations were made.
\item \code{amostra_tipo}. Type of soil sample taken, i.e. simple or composed.
\item \code{amostra_quanti}. Number of soil samples taken.
\item \code{amostra_area}. Sampling area.
}
}
}
\examples{
\dontrun{
res <- observations(dataset = paste("ctb000", 4:9, sep = ""))
str(res)
}
}
\seealso{
\code{\link[febr]{layers}}
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
