% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFEBR.R
\name{readFEBR}
\alias{readFEBR}
\title{Get soil data}
\usage{
readFEBR(
  data.set,
  data.table = c("identificacao", "versionamento", "metadado", "observacao", "camada"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data.set}{Character vector indicating the identification code of one or more data sets.
Use \code{data.set = "all"} to download all data sets.}

\item{data.table}{Character vector indicating one or more data tables, with supported values
\code{"identificacao"}, \code{"versionamento"}, \code{"metadado"}, \code{"observacao"}, and \code{"camada"}.}

\item{verbose}{(optional) Logical value indicating if informative messages should be displayed.
Generally useful to identify issues—please report to \email{febr-forum@googlegroups.com} if
you find any.}

\item{...}{(optional) Arguments passed to \code{\link[utils]{read.table}}.}
}
\value{
A list of data frames with data from the chosen data sets.
}
\description{
Download soil data from one or more data sets published in the Free Brazilian Repository for
Open Soil Data -- FEBR, \url{https://www.pedometria.org/febr/}.
}
\examples{
\donttest{
res <- readFEBR(data.set = "ctb0003")
}
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
