% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readVocabulary.R
\name{readVocabulary}
\alias{readVocabulary}
\title{Get controlled vocabulary}
\usage{
readVocabulary(...)
}
\arguments{
\item{...}{(optional) Arguments passed to \code{\link[data.table:fread]{data.table::fread()}}.}
}
\value{
An object of class \link[data.table:data.table]{data.table::data.table}.
}
\description{
Download the controlled vocabulary used in the FEBR Soil Data Repository,
\url{https://www.pedometria.org/febr/}.
}
\examples{
vocab <- readVocabulary()
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
