% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomy}
\alias{taxonomy}
\title{Soil taxonomy
Extract and process soil taxonomic data from textual soil classification description.}
\usage{
taxonomy(
  text,
  method = "decompose",
  sep = " ",
  pattern = c(", ", " A ", " textura ")
)
}
\arguments{
\item{text}{Character string with soil classification description (in Portuguese).}

\item{method}{Character string defining the string processing method. Options: \code{decompose}.}

\item{sep}{Character string.
Defaults to \code{sep = " "}.}

\item{pattern}{Character string (in Portuguese).
Defaults to \code{pattern = c(", ", " A ", " textura ")}.}
}
\description{
Soil taxonomy
Extract and process soil taxonomic data from textual soil classification description.
}
\examples{
text <-
  c("CAMBISSOLO HÁPLICO Ta Eutrófico léptico A proeminente textura média",
    "PLANOSSOLO HÁPLICO Distrófico solódico êndico plintossólico, textura média, Tb",
    "CHERNOSSOLO ARGILÚVICO Órtico típico textura média com cascalho/argilosa com cascalho",
    "ARGISSOLO VERMELHO-AMARELO",
    "Latossolo"
  )
taxonomy(text)
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
